Seperated interpreter and main into different files
